Removed bug to trigger PTT multiple times Co-authored-by: Daniel Lang <dlang8428@gmail.com>
