--- App/automotive/src/main/java/com/amos/infotaimos/model/AmosVoiceInteractionService.kt
@@ -1,10 +1,0 @@
-package com.amos.infotaimos.model
-
-import android.service.voice.VoiceInteractionService
-
-class AmosVoiceInteractionService: VoiceInteractionService() {
-
-    override fun onReady() {
-        super.onReady()
-    }
-}
